Add the new 'MudExIconPicker' component to updated docs
Added a new component 'MudExIconPicker' to select icons to docs. This component is used in the API overview and in the ComponentGrid as MudExObjectEdit config. Updated the patch version in MudBlazor.Extensions project file to reflect this new addition. Also, updated the COPY timestamps in README.md and added the new component in the CHANGELOG.md for both MudBlazor.Extensions/Docs and MudBlazor.Extensions/wwwroot/docs directories.