Fix JavaScript null exception on dialog title (#28)
This commit resolves a null exception in 'baseDialogExtensionHandler.js' and 'mudBlazorExtensions.all.js'. The issue was occurring when 'dialogTitleEl' was null and we were trying to access innerText property of it. Now, we ensure 'dialogTitleEl' is not null before accessing its property, thus, preventing the null exception. Minor formatting changes have also been made in '.DotSettings.user'.